LayoutTests/platform/glib/imported/w3c/web-platform-tests/content-security-policy/securitypolicyviolation/securitypolicyviolation-block-cross-origin-image-from-script.sub-expected.txt



FAIL Non-redirected cross-origin URLs are not stripped. assert_equals: expected 0 but got 2
